Introducing changes required to run scripts in IDC Inflow image building
Following changes were made to mage scripts applicable for unattended installation.

Adding environment variables:
- OMMIT_REBOOT - when set causes scripts to ommit reboot command
- OMMIT_SERVICE_START - when set prevenst scripts from starting services

Bumping up target kernel version from 5.15.0-57 to 5.15.0-73 - older version was no longer valid.

+++ base/2_kernel_setup.sh
@@ -3,8 +3,8 @@
 export DEBIAN_FRONTEND=noninteractive
 export DEBIAN_FRONTEND=noninteractive
 export NEEDRESTART_MODE=a
-KERNEL_NAME="linux-image-5.15.0-57-generic"
-KERNEL_VERSION="5.15.0-57"
+KERNEL_NAME="linux-image-5.15.0-73-generic"
+KERNEL_VERSION="5.15.0-73"
 
 alias sudo="sudo -E"
 colored_output() {
@@ -50,6 +50,10 @@ else
     colored_output "ERROR: Failed to set kernel ${KERNEL_VERSION}-generic as the default kernel." red
     exit 1
 fi
-colored_output "Rebooting in 10 seconds to apply the new kernel..." blue
-sleep 10
-sudo reboot
+
+if [ -z "$OMMIT_REBOOT" ]
+then
+    colored_output "Rebooting in 10 seconds to apply the new kernel..." blue
+    sleep 10
+    sudo reboot
+fi
