+++ base/9_cleanup.sh
@@ -4,8 +4,8 @@ set -e
 export DEBIAN_FRONTEND=noninteractive
 export DEBIAN_FRONTEND=noninteractive
 export NEEDRESTART_MODE=a
-KERNEL_NAME="linux-image-5.15.0-57-generic"
-KERNEL_VERSION="5.15.0-57"
+KERNEL_NAME="linux-image-5.15.0-73-generic"
+KERNEL_VERSION="5.15.0-73"
 
 alias sudo="sudo -E"
 colored_output() {
@@ -26,6 +26,9 @@ sudo apt update &&\
   sudo apt autoremove
 
 # Reboot
-colored_output "Rebooting the system in 10 seconds..." blue
-sleep 10
-sudo reboot
+if [ -z "$OMMIT_REBOOT" ]
+then
+    colored_output "Rebooting the system in 10 seconds..." blue
+    sleep 10
+    sudo reboot
+fi
